refactor to use shared components
